+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64_x86_64-simulator/FitCloudKit.framework/Headers/FitCloudDataModels.h
@@ -50,6 +50,7 @@
 
 #import <FitCloudKit/FitCloudStorageInfoModel.h>
 #import <FitCloudKit/FitCloudFileInfoModel.h>
+#import <FitCloudKit/FitCloudFileDetailsInfoModel.h>
 
 /// Specific client defines
 #import <FitCloudKit/Huashengda.h>